pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "social-video-automation"
version = "0.1.0"
description = "AI-powered video generation and social media auto-posting for PeterMat brand"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Peter Ball" }
]
keywords = ["video-generation", "social-media", "automation", "ai", "petermat"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",

    # HTTP clients
    "httpx>=0.28.0",
    "aiohttp>=3.11.0",

    # Data validation
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",

    # Task scheduling
    "celery[redis]>=5.4.0",
    "apscheduler>=3.10.0",

    # AI integrations
    "openai>=1.0.0",
    "google-generativeai>=0.8.0",

    # Social media APIs
    "ayrshare>=1.0.0",

    # Video generation
    "creatomate>=1.0.0",

    # Utilities
    "python-dotenv>=1.0.1",
    "structlog>=24.4.0",
    "tenacity>=9.0.0",

    # CLI
    "typer>=0.12.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[project.scripts]
sva = "social_video_automation.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/social_video_automation"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "ARG", "SIM"]
ignore = ["E501", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["social_video_automation"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
